zebra: netlink: add srcdest route support
This implements processing the RTA_SRC attribute on routes received from
kernel, passing along the source prefix to the RIB functions, and adds
the attribute on updates to the kernel.

NB: "src" was in use as variable name for the "prefsrc" attribute on
routes, which indicates the preferred source address for IPv4.  These
variables have been renamed to "prefsrc".

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>